+++ src/httpserver/SubTasksHandler.java
@@ -1,66 +1,65 @@
-package httpServer;
+package httpserver;
 
 import com.google.gson.reflect.TypeToken;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import exceptions.NotFoundException;
 import manager.TaskManager;
-import task.Task;
+import task.SubTask;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Optional;
 
 
-public class TasksHandler extends BaseHttpHandler implements HttpHandler {
+public class SubTasksHandler extends BaseHttpHandler implements HttpHandler {
     TaskManager taskManager;
 
-    public TasksHandler(TaskManager taskManager) {
+    public SubTasksHandler(TaskManager taskManager) {
         this.taskManager = taskManager;
     }
-
     @Override
     public void handle(HttpExchange httpExchange) throws IOException {
         Endpoint endpoint = getEndpoint(httpExchange.getRequestURI().getPath(), httpExchange.getRequestMethod());
 
         switch (endpoint) {
-            case GET_TASKS: {
-                handleGetTasks(httpExchange);
+            case GET_SUBTASKS: {
+                handleGetSubTasks(httpExchange);
                 break;
             }
-            case GET_TASK_BY_ID: {
-                handleGetTaskById(httpExchange);
+            case GET_SUBTASK_BY_ID: {
+                handleGetSubTaskById(httpExchange);
                 break;
             }
-            case POST_TASKS: {
-                handlePostTasks(httpExchange);
+            case POST_SUBTASKS: {
+                handlePostSubTasks(httpExchange);
             }
-            case DELETE_TASK:
-                handlerDeleteTasks(httpExchange);
+            case DELETE_SUBTASK:
+                handlerDeleteSubTasks(httpExchange);
             default:
                 sendNotFound(httpExchange, errToJson(ENDPOINT_UNKNOWN_ERR));
         }
 
     }
 
-    private void handleGetTasks(HttpExchange httpExchange) throws IOException {
+    private void handleGetSubTasks(HttpExchange httpExchange) throws IOException {
         try {
-            ArrayList<Task> tasks = (ArrayList<Task>) taskManager.getTasks();
-            sendText(httpExchange, taskToJson(tasks), 200);
+            ArrayList<SubTask> subTasks = (ArrayList<SubTask>) taskManager.getSubTasks();
+            sendText(httpExchange, taskToJson(subTasks), 200);
         } catch (Exception e) {
             sendText(httpExchange, errToJson(e.getMessage()), 500);
         }
     }
 
-    private void handleGetTaskById(HttpExchange httpExchange) throws IOException {
+    private void handleGetSubTaskById(HttpExchange httpExchange) throws IOException {
         try {
             Optional<Integer> taskIdOpt = getId(httpExchange);
             if (taskIdOpt.isEmpty()) {
                 sendText(httpExchange, errToJson(ILLEGAL_ID_ERR), 400);
                 return;
             }
-            Task task = taskManager.getTaskById(taskIdOpt.get());
-            sendText(httpExchange, taskToJson(task), 200);
+            SubTask subTask = taskManager.getSubTaskById(taskIdOpt.get());
+            sendText(httpExchange, taskToJson(subTask), 200);
         } catch (NotFoundException e) {
             sendNotFound(httpExchange, errToJson(e.getMessage()));
         } catch (Exception e) {
@@ -68,39 +67,41 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
         }
     }
 
-    private void handlePostTasks(HttpExchange httpExchange) throws IOException {
+    private void handlePostSubTasks(HttpExchange httpExchange) throws IOException {
         try {
-            Optional<Task> taskOpt = jsonToTask(
+            Optional<SubTask> taskOpt = jsonToTask(
                     new String(httpExchange.getRequestBody().readAllBytes(), DEFAULT_CHARSET),
-                    new TaskTypeToken().getType());
-            if (taskOpt.isEmpty()) {
+                    new SubTaskTypeToken().getType());
+            if(taskOpt.isEmpty()) {
                 sendText(httpExchange, errToJson(ILLEGAL_JSON_ERR), 400);
                 return;
             }
-            Task task = taskOpt.get();
+            SubTask subTask = taskOpt.get();
 
-            if (task.getTaskId() == null) {
-                taskManager.addTask(task);
-                sendText(httpExchange, successToJson(TASK_ADD), 201);
-            } else {
-                taskManager.updateTask(task);
-                sendText(httpExchange, successToJson(TASK_UPDATED), 201);
+            try {
+                if (subTask.getTaskId() == null) {
+                    taskManager.addSubTask(subTask);
+                    sendText(httpExchange, successToJson(TASK_ADD), 201);
+                } else {
+                    taskManager.updateSubTask(subTask);
+                    sendText(httpExchange, successToJson(TASK_UPDATED), 201);
+                }
+            } catch (IllegalArgumentException e)  {
+                sendHasInteractions(httpExchange, errToJson(e.getMessage()));
             }
-        } catch (IllegalArgumentException e) {
-            sendHasInteractions(httpExchange, errToJson(e.getMessage()));
         } catch (Exception e) {
             sendText(httpExchange, errToJson(e.getMessage()), 500);
         }
     }
 
-    private void handlerDeleteTasks(HttpExchange httpExchange) throws IOException {
+    private void handlerDeleteSubTasks(HttpExchange httpExchange) throws IOException {
         try {
             Optional<Integer> taskIdOpt = getId(httpExchange);
-            if (taskIdOpt.isEmpty()) {
+            if(taskIdOpt.isEmpty()) {
                 sendText(httpExchange, errToJson(ILLEGAL_ID_ERR), 400);
                 return;
             }
-            taskManager.deleteTaskById(taskIdOpt.get());
+            taskManager.deleteSubTaskById(taskIdOpt.get());
             sendText(httpExchange, successToJson(TASK_DELETED), 200);
         } catch (NotFoundException e) {
             sendNotFound(httpExchange, errToJson(e.getMessage()));
@@ -110,5 +111,5 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
     }
 }
 
-class TaskTypeToken extends TypeToken<Task> {
+class SubTaskTypeToken extends TypeToken<SubTask> {
 }
