+++ test/httpserver/HttpTaskManagerEpicsTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver;
 
 import com.google.gson.reflect.TypeToken;
 import manager.InMemoryTaskManager;
@@ -15,11 +15,8 @@ import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.time.Duration;
-import java.time.LocalDateTime;
 import java.util.List;
 
-import static httpServer.BaseHttpHandler.TASK_UPDATED;
 import static manager.InMemoryTaskManager.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static util.TestConstants.*;
