+++ test/httpserver/HttpTaskManagerPrioritizedTest.java
@@ -1,6 +1,5 @@
-package httpServer;
+package httpserver;
 
-import com.google.gson.reflect.TypeToken;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
 import org.junit.jupiter.api.AfterEach;
@@ -20,13 +19,11 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static httpServer.BaseHttpHandler.TASK_UPDATED;
-import static manager.InMemoryTaskManager.TASK_CROSS_ERROR;
-import static manager.InMemoryTaskManager.TASK_NOT_FOUND_ERR;
-import static org.junit.jupiter.api.Assertions.*;
 import static util.TestConstants.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class HttpTaskManagerHistoryTest {
+public class HttpTaskManagerPrioritizedTest {
 
     // создаём экземпляр InMemoryTaskManager
     TaskManager manager = new InMemoryTaskManager();
@@ -34,7 +31,7 @@ public class HttpTaskManagerHistoryTest {
     HttpTaskServer taskServer = new HttpTaskServer(manager);
     BaseHttpHandler baseHttpHandler= new BaseHttpHandler();
 
-    public HttpTaskManagerHistoryTest() {
+    public HttpTaskManagerPrioritizedTest() {
     }
 
     @BeforeEach
@@ -50,22 +47,19 @@ public class HttpTaskManagerHistoryTest {
     }
 
     @Test
-    public void testGetHistory() throws IOException, InterruptedException {
-        Task task = new Task("NEW_TASK_NAME", NEW_TASK_DESC,
+    public void testGetPrioritizedTask() throws IOException, InterruptedException {
+        Task task = new Task(NEW_TASK_NAME, NEW_TASK_DESC,
                 TaskStatus.NEW, LocalDateTime.now(),Duration.ofMinutes(5));
         int taskId = manager.addTask(task).getTaskId();
         Epic epic = new Epic(NEW_EPIC_NAME, NEW_EPIC_DESC, TaskStatus.NEW);
         int epicId = manager.addEpic(epic).getTaskId();
-        SubTask subTask = new SubTask("NEW_SUBTASK_NAME", NEW_SUBTASK_DESC,
+        SubTask subTask = new SubTask(NEW_SUBTASK_NAME, NEW_SUBTASK_DESC,
                 TaskStatus.NEW, epicId, LocalDateTime.now().plusHours(1),Duration.ofMinutes(5));
         int subTaskId = manager.addSubTask(subTask).getTaskId();
 
-        manager.getTaskById(taskId);
-        manager.getEpicById(epicId);
-
         // создаём HTTP-клиент и запрос
         HttpClient client = HttpClient.newHttpClient();
-        URI url = URI.create("http://localhost:8080/history");
+        URI url = URI.create("http://localhost:8080/prioritized");
         HttpRequest request = HttpRequest.newBuilder().uri(url).GET().build();
 
         HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
@@ -76,10 +70,10 @@ public class HttpTaskManagerHistoryTest {
         List<Task> tasksListFromHttpServer = (List<Task>) baseHttpHandler.jsonToTask(response.body(),
                 new ListTaskTypeToken().getType()).get();
 
+        System.out.println(tasksListFromHttpServer);
         assertNotNull(tasksListFromHttpServer, TASK_NOT_RETURN_ERR);
-        assertEquals(3, tasksListFromHttpServer.size(), INCORRECT_NUM_OF_TASK_ERR);
-        assertEquals("NEW_TASK_NAME", tasksListFromHttpServer.get(0).getName(), INCORRECT_TASK_NAME_ERR);
-        assertEquals(NEW_EPIC_NAME, tasksListFromHttpServer.get(1).getName(), "Некорректное имя эпика");
-        assertEquals("NEW_SUBTASK_NAME", tasksListFromHttpServer.get(2).getName(), INCORRECT_SUBTASK_NAME_ERR);
+        assertEquals(2, tasksListFromHttpServer.size(), INCORRECT_NUM_OF_TASK_ERR);
+        assertEquals(NEW_TASK_NAME, tasksListFromHttpServer.get(0).getName(), INCORRECT_TASK_NAME_ERR);
+        assertEquals(NEW_SUBTASK_NAME, tasksListFromHttpServer.get(1).getName(), INCORRECT_SUBTASK_NAME_ERR);
     }
 }
